added userId i

+++ src/blogs/blogs.service.ts
@@ -1,5 +1,5 @@
 import { PostsRepository } from './../posts/posts.repository';
-import { Injectable } from '@nestjs/common';
+import { Injectable, ForbiddenException } from '@nestjs/common';
 import { CreateBlogDto } from './dto/create-blog.dto';
 import { BlogDocument } from './schemas/blog.schema';
 import { BlogsRepository } from './blogs.repository';
@@ -37,11 +37,23 @@ export class BlogsService {
     };
   }
 
-  async updateBlog(blog: UpdateBlogDto, id: string) {
+  async updateBlog(blog: UpdateBlogDto, id: string, userId: string) {
+    const foundBlog = await this.blogsRepository.findBlogWithOwnerInfo(id);
+
+    if (foundBlog.blogOwnerInfo.userId != userId) {
+      throw new ForbiddenException();
+    }
+
     return this.blogsRepository.updateBlog(blog, id);
   }
 
-  async removeBlog(id: string): Promise<boolean> {
+  async removeBlog(id: string, userId: string): Promise<boolean> {
+    const foundBlog = await this.blogsRepository.findBlogWithOwnerInfo(id);
+
+    if (foundBlog.blogOwnerInfo.userId != userId) {
+      throw new ForbiddenException();
+    }
+
     const isDeleted = await this.blogsRepository.removeBlog(id);
     if (isDeleted) {
       this.postsRepository.removePostByBlogId(id);
